fixed calculations in SetMapMode() (use FP division, not integer)
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@20867 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/msw/dc.cpp
@@ -1615,8 +1615,8 @@ void wxDC::SetMapMode(int mode)
             return;
         }
 
-        double mm2pixelsX = pixel_width / mm_width,
-               mm2pixelsY = pixel_height / mm_height;
+        double mm2pixelsX = (double)pixel_width / mm_width,
+               mm2pixelsY = (double)pixel_height / mm_height;
 
         switch (mode)
         {
